Make new shit

+++ Assets/Source/Spell/SpellSpawnType.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public abstract class SpellSpawnType : MonoBehaviour
 {
     public abstract int Count { get; set; }
-    public abstract int Size { get; set; }
+    public abstract float Size { get; set; }
     public abstract Action PerformSpawn(Spell spell);
     public abstract void PerformModificator(Spell spell);
